+++ README.md
@@ -12,6 +12,8 @@ Usage examples: https://github.com/NREL/buildstock-query/tree/main/example_usage
 
 A standalone CLI utility to scan an Athena database and identify (or drop) stale tables and views whose underlying S3 data no longer exists.
 
+Programmatic usage examples: [example_usage/aws_athena_tools.ipynb](example_usage/aws_athena_tools.ipynb)
+
 ### Usage
 
 ```bash
@@ -31,23 +33,9 @@ aws_athena_cleanup --database my_db --workgroup primary --s3-output s3://my-buck
 aws_athena_cleanup --help
 ```
 
-### Programmatic Usage
-
-```python
-from buildstock_query.tools import aws_athena_cleanup
-
-summary = aws_athena_cleanup(
-    database="my_db",
-    workgroup="primary",
-    region="us-west-2",
-    drop=False,  # set True to actually remove stale objects
-)
-print(summary["stale_tables"])
-```
-
 ## AWS Athena Table Search Tool
 
-A standalone CLI utility to search for a table across all Glue/Athena databases in your catalog.
+A standalone CLI utility to search for a table across all Glue/Athena databases in your catalog. See programmatic usage in example_usage/aws_athena_tools.ipynb.
 
 ### Usage
 
@@ -69,20 +57,5 @@ aws_athena_table_search --table my_table --region us-east-1
 
 # Help: see available configuration setting
 aws_athena_table_search --help
-```
-
-### Programmatic Usage
-
-```python
-from buildstock_query.tools import aws_athena_table_search
-
-matches = aws_athena_table_search(
-    table_name="my_table",
-    region="us-west-2",
-    substring=True,  # substring match
-    database_filter="resstock",  # optional: limit to certain databases
-)
-for m in matches:
-    print(f"{m['database']}.{m['table']} -> {m['s3_location']}")
 ```
